Workaround for github-flavored markdown not ignoring newlines within paragraphs
I use sphinx-markdown-builder to generate release notes for GitHub
releases.

Unlike CommonMark, GitHub's flavor of Markdown interprets input like
"line1\nline2" as "line1<br/>line2".

As a workaround, I could write each paragraph in my *.rst sources in
only one line *.rst.  But I'm not sure I want to do that; it would
be surprising to people who work on the *.rst sources because all
other builders (HTML, man, markdown+commonmark) handle this just fine.

Add a workaround to produce a faithful version GitHub-flavored
markdown by joining all multiline paragraphs into a single line.

I'm not aware of significant downsides yet (except that the Markdown
is uglier) but I haven't tested beyond my real-world use.

To be safe, I put it behind a config option, if only to not muddy
the waters as to which dialect of Markdown we are targeting.

Might as well make that option open for extensions, because there
are quite a few flavors:

	$ pandoc --list-input-formats | grep mark -i | paste -sd ,
	commonmark,commonmark_x,markdown,markdown_github,markdown_mmd,markdown_phpextra,markdown_strict